backend/requirements.txt
# 프레임워크
fastapi>=0.110
uvicorn[standard]>=0.27
websockets>=12.0

# DB
sqlalchemy[asyncio]>=2.0
aiosqlite>=0.19,<0.22
alembic>=1.13

# 설정
pydantic-settings>=2.1

# 데이터 수집
yfinance>=0.2.36
ccxt>=4.2
pykrx>=1.0.45
finance-datareader>=0.9.66

# 한국투자증권
python-kis>=2.1

# 지표 계산
pandas>=2.2
numpy>=1.26
pandas-ta-classic>=0.1.0

# 스케줄러
apscheduler>=3.10,<4.0

# 텔레그램
python-telegram-bot>=20.7

# 서머타임
pytz>=2024.1

# HTTP
httpx>=0.27

# 로깅
loguru>=0.7

# 개발
pytest>=8.0
pytest-asyncio>=0.23
ruff>=0.3
